+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerWithSeconds.test.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import DateTimePickerExampleTemplate from '../../common/DateTimePickerExampleTemplate';
-
-const DateTimePickerWithSeconds = () => (
-  <DateTimePickerExampleTemplate showSeconds />
-);
-
-export default DateTimePickerWithSeconds;