+++ packages/terra-framework-docs/src/terra-dev-site/test/date-picker/DatePickerOnChangeRaw.test.jsx
@@ -1,0 +1,59 @@
+import React, { useState } from 'react';
+import classNames from 'classnames/bind';
+import DatePicker from 'terra-date-picker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
+
+const DatePickerOnChangeRaw = () => {
+  const [date, setDate] = useState('');
+  const [iSO, setISO] = useState('');
+  const [inputValue, setInputValue] = useState('');
+  const [isCompleteValue, setIsCompleteValue] = useState('No');
+  const [isValidValue, setIsValidValue] = useState('Yes');
+
+  const handleDateChangeRaw = (event, dateValue, options) => {
+    setDate(dateValue);
+    setISO(options.iSO);
+    setInputValue(options.inputValue);
+    setIsCompleteValue(options.isCompleteValue ? 'Yes' : 'No');
+    setIsValidValue(options.isValidValue ? 'Yes' : 'No');
+  };
+
+  return (
+    <div className={cx('content-wrapper')}>
+      <h3>
+        Selected Date:
+        {' '}
+        <span id="selected-date">{date}</span>
+        <br />
+        <br />
+        ISO String:
+        {' '}
+        <span id="iso">{iSO}</span>
+        <br />
+        <br />
+        Input Value:
+        {' '}
+        <span id="input-value">{inputValue}</span>
+        <br />
+        <br />
+        Is Date Complete?
+        {' '}
+        <span id="complete-date">{isCompleteValue}</span>
+        <br />
+        <br />
+        Is Date Valid?
+        {' '}
+        <span id="valid-date">{isValidValue}</span>
+      </h3>
+      <DatePicker
+        name="date-input-onchangeraw"
+        onChangeRaw={handleDateChangeRaw}
+        excludeDates={['2019-04-01', '2019-04-02']}
+      />
+    </div>
+  );
+};
+
+export default DatePickerOnChangeRaw;