+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrate docs to terra framework docs
+
 ## 6.60.0 - (November 16, 2021)
 
 * Changed